+++ jest.config.js
@@ -4,4 +4,14 @@ module.exports = {
   testEnvironment: "node",
   resetMocks: false,
   setupFiles: ["jest-localstorage-mock"],
+  collectCoverage: true,
+  coverageReporters: ["text", "html", "cobertura"],
+  coverageThreshold: {
+    global: {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
+    },
+  },
 };