+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### dependabot: \#49 Bump cross-spawn from 7.0.3 to 7.0.6
 
+### Added
+
+- Code coverage with 100% threshold
+
 ## [1.2.0] - 2025-01-13
 
 ### Added